+++ system/apps/guide_board_lifecycle/lib/guide_board_lifecycle/supervisor.ex
@@ -12,6 +12,8 @@ defmodule GuideBoardLifecycle.Supervisor do
   use Supervisor
 
   alias GuideBoardLifecycle.DrawStroke.AnswerDrawStrokeRequestsStarter
+  alias GuideBoardLifecycle.LeaveBoard.AnswerLeaveBoardRequestsStarter
+  alias GuideBoardLifecycle.LeaveBoard.PeerDepartedV1ToMesh
   alias GuideBoardLifecycle.StrokeDrawnV1ToMesh.StrokeDrawnV1ToMesh
 
   def start_link, do: Supervisor.start_link(__MODULE__, [], name: __MODULE__)
@@ -20,9 +22,11 @@ defmodule GuideBoardLifecycle.Supervisor do
   def init([]) do
     children = [
       handler(StrokeDrawnV1ToMesh),
+      handler(PeerDepartedV1ToMesh),
       {DynamicSupervisor,
        name: GuideBoardLifecycle.MeshSubscriberSupervisor, strategy: :one_for_one},
-      AnswerDrawStrokeRequestsStarter
+      AnswerDrawStrokeRequestsStarter,
+      AnswerLeaveBoardRequestsStarter
     ]
 
     Supervisor.init(children, strategy: :one_for_one)